requirements.txt
Django==1.8.2
Pillow==2.8.1
argparse==1.2.1
django-extensions==1.5.5
djangorestframework==3.1.3
psycopg2==2.5.4
six==1.9.0
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
